feat: update design

+++ forum/views.py
@@ -1,4 +1,4 @@
-from django.http import HttpResponse, JsonResponse
+from django.http import HttpResponse, JsonResponse, HttpResponseNotFound
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from forum.models import Forum, Balasan
@@ -34,7 +34,7 @@ def create_forum(request, id_restoran):
     form = ForumForm(request.POST or None)
 
     if request.user.peran != 'pengulas':
-        return redirect('forum:show_forum', id_restoran=restoran.id)
+        return HttpResponseNotFound()
 
     if form.is_valid() and request.method == "POST":
         topik = form.save(commit=False)
@@ -56,6 +56,9 @@ def edit_forum(request, id_restoran, id_forum):
     form = ForumForm(request.POST, instance=forum)
     restoran = Restoran.objects.get(pk=id_restoran)
 
+    if request.user.peran != 'pengulas' or request.user.id != forum.user.id:
+        return HttpResponseNotFound()
+
     if request.method == "POST":
         if form.is_valid():
             form.save()
@@ -70,6 +73,9 @@ def delete_forum(request, id_restoran, id_forum):
     forum = Forum.objects.get(pk=id_forum)
     restoran = Restoran.objects.get(pk=id_restoran)
 
+    if request.user.peran != 'pengulas' or request.user.id != forum.user.id:
+        return HttpResponseNotFound()
+
     forum.delete()
     return redirect('forum:show_forum', id_restoran=restoran.id)
 
@@ -81,8 +87,8 @@ def edit_balasan(request, id_restoran, id_forum, id_balasan):
     balasan = Balasan.objects.get(pk=id_balasan)
     form = BalasanForm(request.POST, instance=balasan)
 
-    if request.user.peran != 'pengulas':
-        return redirect('forum:show_forum_by_id', id_restoran=restoran.id, id_forum=forum.id)
+    if request.user.peran != 'pengulas' or request.user != balasan.user:
+        return HttpResponseNotFound()
     
     if request.method == "POST":
         if form.is_valid():
@@ -99,8 +105,8 @@ def delete_balasan(request, id_restoran, id_forum, id_balasan):
     restoran = Restoran.objects.get(pk=id_restoran)
     balasan = Balasan.objects.get(pk=id_balasan)
 
-    if request.user.peran != 'pengulas':
-        return redirect('forum:show_forum_by_id', id_restoran=restoran.id, id_forum=forum.id)
+    if request.user.peran != 'pengulas' or request.user != balasan.user:
+        return HttpResponseNotFound()
 
     balasan.delete()
     return redirect('forum:show_forum_by_id', id_restoran=restoran.id, id_forum=forum.id)
@@ -108,7 +114,7 @@ def delete_balasan(request, id_restoran, id_forum, id_balasan):
 
 @require_POST
 @login_required(login_url="login")
-def balas_ajax(request, id_restoran, id_forum):
+def balas(request, id_restoran, id_forum):
     pesan = request.POST.get("pesan")
     user = request.user
     forum = Forum.objects.get(pk=id_forum)
@@ -121,7 +127,7 @@ def balas_ajax(request, id_restoran, id_forum):
 
 def show_balasan(request, id_restoran, id_forum):
     try:
-        data = Balasan.objects.filter(forum=id_forum)
+        data = Balasan.objects.filter(forum=id_forum).order_by("tanggal_posting")
         balasan = []
         for item in data:
             balasan.append({
